Change the devstat generation number from an int to a long.  The int-sized generation was causing unaligned access faults on the Alpha.
I have incremented the devstat version number, since this is an interface
change.  You'll need to recompile libdevstat, systat, iostat, vmstat and
rpc.rstatd along with your kernel.

Partially Submitted by:	Andrew Gallatin <gallatin@cs.duke.edu>

+++ usr.bin/systat/extern.h
@@ -61,11 +61,11 @@ extern int	verbose;
 struct inpcb;
 
 extern struct device_selection *dev_select;
-extern int			generation;
+extern long			generation;
 extern int			num_devices;
 extern int			num_selected;
 extern int			num_selections;
-extern int			select_generation;
+extern long			select_generation;
 
 int	 checkhost __P((struct inpcb *));
 int	 checkport __P((struct inpcb *));
